Core/Misc: More warning fixes

+++ src/server/game/Entities/Object/Object.cpp
@@ -2915,7 +2915,7 @@ void WorldObject::RebuildTerrainSwaps()
         {
             // only add terrain swaps for current map
             MapEntry const* mapEntry = sMapStore.LookupEntry(swap);
-            if (!mapEntry || mapEntry->rootPhaseMap != GetMapId())
+            if (!mapEntry || mapEntry->rootPhaseMap != int32(GetMapId()))
                 continue;
 
             conditions = sConditionMgr->GetConditionsForNotGroupedEntry(CONDITION_SOURCE_TYPE_TERRAIN_SWAP, swap);
